Add salary reprocess crawler command
